[CDEC-371] Remove unecessary `core-test` dependencies
When `Test.Pos.Core.Gen` was added to `cardano-sl-core-test`,
dependencies on `cardano-sl-delegation`, `cardano-sl-block`, and
`cardano-sl-ssc` were introduced. This introduced potential dependency
cycles.

In this commit I push code from block/, delegation/, and ssc/ into core/.
The modules I moved seem to fit fairly well in core/, conceptually, so
the change seems necessary and fitting.
