+++ block/cardano-sl-block.cabal
@@ -14,7 +14,6 @@ cabal-version:       >=1.10
 
 library
   exposed-modules:
-                        Pos.Block.Base
                         Pos.Block.BHelpers
                         Pos.Block.BListener
                         Pos.Block.BlockWorkMode
