+++ core/src/Pos/Core/Block.hs
@@ -4,9 +4,11 @@ module Pos.Core.Block
        , module Pos.Core.Block.Main
        , module Pos.Core.Block.Genesis
        , module Pos.Core.Block.Blockchain
+       , module Pos.Core.Block.Constructors
        ) where
 
-import          Pos.Core.Block.Blockchain
-import          Pos.Core.Block.Genesis
-import          Pos.Core.Block.Main
-import          Pos.Core.Block.Union
+import           Pos.Core.Block.Blockchain
+import           Pos.Core.Block.Constructors
+import           Pos.Core.Block.Genesis
+import           Pos.Core.Block.Main
+import           Pos.Core.Block.Union
