+++ core/src/Pos/Core/Block/Constructors.hs
@@ -1,11 +1,10 @@
 -- | Block constructors and basic functions.
 
-module Pos.Block.Base
+module Pos.Core.Block.Constructors
        ( mkMainBlockExplicit
        , mkMainBlock
        , mkMainHeaderExplicit
        , mkMainHeader
-       , emptyMainBody
 
        , mkGenesisHeader
        , mkGenesisBlock
@@ -14,24 +13,24 @@ module Pos.Block.Base
 
 import           Universum
 
-import           Data.Default (Default (def))
-
-import           Pos.Block.BHelpers ()
-import           Pos.Core (BlockVersion, ChainDifficulty, EpochIndex, GenesisHash (..),
-                           HasDifficulty (..), HasProtocolConstants, HeaderHash, LocalSlotIndex,
-                           SlotId, SlotLeaders, SoftwareVersion, headerHash)
-import           Pos.Core.Block (BlockHeader, BlockSignature (..), GenericBlock (..), GenesisBlock,
-                                 GenesisBlockHeader, GenesisBody (..), GenesisConsensusData (..),
-                                 GenesisExtraBodyData (..), GenesisExtraHeaderData (..), MainBlock,
-                                 MainBlockHeader, MainBody (..), MainConsensusData (..),
-                                 MainExtraBodyData (..), MainExtraHeaderData (..), MainToSign (..),
-                                 mkGenericHeader)
+import           Pos.Binary.Core.Blockchain () -- Bi instances
+import           Pos.Core.Block.Blockchain (GenericBlock (..), mkGenericHeader)
+import           Pos.Core.Block.Genesis (GenesisBody (..), GenesisConsensusData (..),
+                                         GenesisExtraBodyData (..), GenesisExtraHeaderData (..))
+import           Pos.Core.Block.Main (MainBody (..), MainExtraBodyData (..),
+                                      MainExtraHeaderData (..))
+import           Pos.Core.Block.Union (BlockHeader, BlockSignature (..), GenesisBlock,
+                                       GenesisBlockHeader, HeaderHash, MainBlock,
+                                       MainBlockHeader, MainConsensusData (..), MainToSign (..),
+                                       headerHash)
+import           Pos.Core.Common (ChainDifficulty, HasDifficulty (..), SlotLeaders)
+import           Pos.Core.Configuration (GenesisHash (..))
+import           Pos.Core.Delegation.HeavyDlgIndex (ProxySKBlockInfo)
+import           Pos.Core.Slotting (EpochIndex, SlotId)
+import           Pos.Core.Update (BlockVersion, SoftwareVersion)
 import           Pos.Crypto (ProtocolMagic, SecretKey, SignTag (..), hash, proxySign, sign,
                              toPublic)
 import           Pos.Data.Attributes (mkAttributes)
-import           Pos.Delegation.Types (ProxySKBlockInfo)
-import           Pos.Ssc.Base (defaultSscPayload)
-import           Pos.Txp.Base (emptyTxPayload)
 
 ----------------------------------------------------------------------------
 -- Main smart constructors
@@ -132,6 +131,21 @@ mkMainBlockExplicit pm bv sv prevHash difficulty slotId sk pske body =
             (mkAttributes ())
             (hash extraB)
 
+
+{-
+
+# mhueschen: this is not used anywhere in the codebase
+
+# removed from exports
+--        , emptyMainBody
+
+# removed from imports
+-- import           Data.Default (Default (def))
+-- import           Pos.Ssc.Base (defaultSscPayload)
+-- import           Pos.Txp.Base (emptyTxPayload)
+-- import           Pos.Core.Slotting (LocalSlotIndex)
+-- import           Pos.Core.Configuration (HasProtocolConstants)
+
 -- | Empty (i. e. no payload) body of main block for given local slot index.
 emptyMainBody
     :: HasProtocolConstants
@@ -144,6 +158,7 @@ emptyMainBody slot =
     , _mbDlgPayload = def
     , _mbUpdatePayload = def
     }
+-}
 
 ----------------------------------------------------------------------------
 -- Genesis smart constructors
