+++ core/src/Pos/Core/Delegation/HeavyDlgIndex.hs
@@ -2,6 +2,7 @@ module Pos.Core.Delegation.HeavyDlgIndex
        ( HeavyDlgIndex (..)
        , ProxySigHeavy
        , ProxySKHeavy
+       , ProxySKBlockInfo
        ) where
 
 import           Universum
@@ -10,7 +11,7 @@ import qualified Data.Text.Buildable
 import           Formatting (bprint, build)
 
 import           Pos.Core.Slotting (EpochIndex)
-import           Pos.Crypto (ProxySecretKey (..), ProxySignature)
+import           Pos.Crypto (ProxySecretKey (..), ProxySignature, PublicKey)
 
 -- | Witness for heavy delegation signature -- epoch in which
 -- certificate starts being active. It is needed for replay attack
@@ -31,3 +32,8 @@ type ProxySigHeavy a = ProxySignature HeavyDlgIndex a
 
 -- | Heavy delegation PSK.
 type ProxySKHeavy = ProxySecretKey HeavyDlgIndex
+
+-- | Heavyweight PSK with real leader public key (because heavyweight
+-- psks have redelegation feature, so pskIssuerPk hPsk /= leader in
+-- general case). This is used to create a block header only.
+type ProxySKBlockInfo = Maybe (ProxySKHeavy, PublicKey)
