+++ core/src/Pos/Core/Ssc.hs
@@ -3,6 +3,7 @@ module Pos.Core.Ssc
        , module Pos.Core.Ssc.CommitmentsMap
        , module Pos.Core.Ssc.Opening
        , module Pos.Core.Ssc.OpeningsMap
+       , module Pos.Core.Ssc.CommitmentAndOpening
        , module Pos.Core.Ssc.Payload
        , module Pos.Core.Ssc.Proof
        , module Pos.Core.Ssc.SharesDistribution
@@ -15,6 +16,7 @@ import           Pos.Core.Ssc.Commitment
 import           Pos.Core.Ssc.CommitmentsMap
 import           Pos.Core.Ssc.Opening
 import           Pos.Core.Ssc.OpeningsMap
+import           Pos.Core.Ssc.CommitmentAndOpening
 import           Pos.Core.Ssc.Payload
 import           Pos.Core.Ssc.Proof
 import           Pos.Core.Ssc.SharesDistribution
